refactor: extract overview token utilities
